renamed fonts, rebuilded and applied fonts
