+++ lineQueueClient/android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.warehouseclient",
+    package = "com.linequeueclient",
 )
 
 android_resource(
     name = "res",
-    package = "com.warehouseclient",
+    package = "com.linequeueclient",
     res = "src/main/res",
 )
