+++ lineQueueClient/android/app/src/main/java/com/linequeueclient/MainApplication.java
@@ -1,4 +1,4 @@
-package com.warehouseclient;
+package com.linequeueclient;
 
 import android.app.Application;
