+++ lineQueueClient/android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'warehouseClient'
+rootProject.name = 'lineQueueClient'
 
 include ':app'
