+++ lineQueueClient/ios/lineQueueClient.xcodeproj/xcshareddata/xcschemes/lineQueueClient-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "warehouseClient-tvOS.app"
-               BlueprintName = "warehouseClient-tvOS"
-               ReferencedContainer = "container:warehouseClient.xcodeproj">
+               BuildableName = "lineQueueClient-tvOS.app"
+               BlueprintName = "lineQueueClient-tvOS"
+               ReferencedContainer = "container:lineQueueClient.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "warehouseClient-tvOSTests.xctest"
-               BlueprintName = "warehouseClient-tvOSTests"
-               ReferencedContainer = "container:warehouseClient.xcodeproj">
+               BuildableName = "lineQueueClient-tvOSTests.xctest"
+               BlueprintName = "lineQueueClient-tvOSTests"
+               ReferencedContainer = "container:lineQueueClient.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "warehouseClient-tvOSTests.xctest"
-               BlueprintName = "warehouseClient-tvOSTests"
-               ReferencedContainer = "container:warehouseClient.xcodeproj">
+               BuildableName = "lineQueueClient-tvOSTests.xctest"
+               BlueprintName = "lineQueueClient-tvOSTests"
+               ReferencedContainer = "container:lineQueueClient.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "warehouseClient-tvOS.app"
-            BlueprintName = "warehouseClient-tvOS"
-            ReferencedContainer = "container:warehouseClient.xcodeproj">
+            BuildableName = "lineQueueClient-tvOS.app"
+            BlueprintName = "lineQueueClient-tvOS"
+            ReferencedContainer = "container:lineQueueClient.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "warehouseClient-tvOS.app"
-            BlueprintName = "warehouseClient-tvOS"
-            ReferencedContainer = "container:warehouseClient.xcodeproj">
+            BuildableName = "lineQueueClient-tvOS.app"
+            BlueprintName = "lineQueueClient-tvOS"
+            ReferencedContainer = "container:lineQueueClient.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "warehouseClient-tvOS.app"
-            BlueprintName = "warehouseClient-tvOS"
-            ReferencedContainer = "container:warehouseClient.xcodeproj">
+            BuildableName = "lineQueueClient-tvOS.app"
+            BlueprintName = "lineQueueClient-tvOS"
+            ReferencedContainer = "container:lineQueueClient.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
