+++ lineQueueClient/ios/lineQueueClient.xcodeproj/xcshareddata/xcschemes/lineQueueClient.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "warehouseClient.app"
-               BlueprintName = "warehouseClient"
-               ReferencedContainer = "container:warehouseClient.xcodeproj">
+               BuildableName = "lineQueueClient.app"
+               BlueprintName = "lineQueueClient"
+               ReferencedContainer = "container:lineQueueClient.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "warehouseClientTests.xctest"
-               BlueprintName = "warehouseClientTests"
-               ReferencedContainer = "container:warehouseClient.xcodeproj">
+               BuildableName = "lineQueueClientTests.xctest"
+               BlueprintName = "lineQueueClientTests"
+               ReferencedContainer = "container:lineQueueClient.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "warehouseClientTests.xctest"
-               BlueprintName = "warehouseClientTests"
-               ReferencedContainer = "container:warehouseClient.xcodeproj">
+               BuildableName = "lineQueueClientTests.xctest"
+               BlueprintName = "lineQueueClientTests"
+               ReferencedContainer = "container:lineQueueClient.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "warehouseClient.app"
-            BlueprintName = "warehouseClient"
-            ReferencedContainer = "container:warehouseClient.xcodeproj">
+            BuildableName = "lineQueueClient.app"
+            BlueprintName = "lineQueueClient"
+            ReferencedContainer = "container:lineQueueClient.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "warehouseClient.app"
-            BlueprintName = "warehouseClient"
-            ReferencedContainer = "container:warehouseClient.xcodeproj">
+            BuildableName = "lineQueueClient.app"
+            BlueprintName = "lineQueueClient"
+            ReferencedContainer = "container:lineQueueClient.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "warehouseClient.app"
-            BlueprintName = "warehouseClient"
-            ReferencedContainer = "container:warehouseClient.xcodeproj">
+            BuildableName = "lineQueueClient.app"
+            BlueprintName = "lineQueueClient"
+            ReferencedContainer = "container:lineQueueClient.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
