+++ lineQueueClient/ios/lineQueueClient/AppDelegate.h
