+++ lineQueueClient/ios/lineQueueClient/Info.plist
@@ -5,7 +5,7 @@
 	<key>CFBundleDevelopmentRegion</key>
 	<string>en</string>
 	<key>CFBundleDisplayName</key>
-	<string>warehouseClient</string>
+	<string>lineQueueClient</string>
 	<key>CFBundleExecutable</key>
 	<string>$(EXECUTABLE_NAME)</string>
 	<key>CFBundleIdentifier</key>
@@ -39,9 +39,8 @@
 	<key>UIViewControllerBasedStatusBarAppearance</key>
 	<false/>
 	<key>NSLocationWhenInUseUsageDescription</key>
-	<string></string>
+	<string/>
 	<key>NSAppTransportSecurity</key>
-	<!--See http://ste.vn/2015/06/10/configuring-app-transport-security-ios-9-osx-10-11/ -->
 	<dict>
 		<key>NSExceptionDomains</key>
 		<dict>
@@ -52,5 +51,18 @@
 			</dict>
 		</dict>
 	</dict>
+	<key>UIAppFonts</key>
+	<array>
+		<string>Alef Bold.ttf</string>
+		<string>Alef Regular.ttf</string>
+		<string>Assistant-Bold.ttf</string>
+		<string>Assistant-ExtraBold.ttf</string>
+		<string>Assistant-Regular.ttf</string>
+		<string>Assistant-SemiBold.ttf</string>
+		<string>Assistant Bold.ttf</string>
+		<string>Assistant ExtraBold.ttf</string>
+		<string>Assistant Regular.ttf</string>
+		<string>Assistant SemiBold.ttf</string>
+	</array>
 </dict>
 </plist>
