+++ lineQueueClient/ios/lineQueueClientTests/lineQueueClientTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface warehouseClientTests : XCTestCase
+@interface lineQueueClientTests : XCTestCase
 
 @end
 
-@implementation warehouseClientTests
+@implementation lineQueueClientTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
