+++ lineQueueClient/src/fonts/Assistant Bold.ttf
