+++ lineQueueClient/src/fonts/Assistant SemiBold.ttf
